chore: link reviews to db
