+++ services/contents/src/schema/types.ts
@@ -1,15 +1,15 @@
-import {gql} from "apollo-server";
+import { gql } from 'apollo-server';
 
 const typeDefs = gql`
-    type Query {
-        contents: [Content]
-    }
+  type Query {
+    contents: [Content]
+  }
 
-    type Content @key(fields: "id") {
-        id: ID!
-        title: String
-        year: Int
-    }
+  type Content @key(fields: "id") {
+    id: ID!
+    title: String
+    year: Int
+  }
 `;
 
 export { typeDefs };
