+++ services/reviews/src/schema/resolvers.ts
@@ -3,15 +3,16 @@ import { Context } from '../context';
 // TODO: Probably will need codegen for resolve types (any): https://the-guild.dev/graphql/codegen/plugins/typescript/typescript-resolvers
 const resolvers = {
   Query: {
+    // TODO: Comments?
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    contents: (_parent: any, _args: any, ctx: Context) => ctx.prisma.content.findMany(),
+    reviews: (_parent: any, _args: any, ctx: Context) => ctx.prisma.review.findMany(),
   },
   Content: {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    __resolveReference: (content: any, _args: any, ctx: Context) =>
-      ctx.prisma.content.findUnique({
+    reviews: (content: any, _args: any, ctx: Context) =>
+      ctx.prisma.review.findMany({
         where: {
-          id: content.id,
+          contentId: content.id,
         },
       }),
   },
