+++ services/reviews/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "@acme/tsconfig/node.json",
-  "include": ["."],
+  "include": ["src/**/*"],
   "exclude": ["dist", "build", "node_modules"]
 }
